+++ docs/realm.js
@@ -307,18 +307,27 @@ class Realm {
   compact() {}
 
   /**
-   * Writes a compacted copy of the Realm to the given path.
+   * Writes a compacted copy of the Realm α) to the given path or β) with the given output
+   *
+   * For invocation with α):
+   *   * Input Realms may be local or synced, encrypted or non-encrypted
+   *   * Output Realms will be local only, encrypted or non-encrypted
+   * Deprecation Warning: Invoking `writeCopyTo` with a path string is deprecated and will be removed in the next breaking release.
+   * Please invoke `writeCopyTo` with a {@link Realm~Configuration | Configuration}.
+   *
+   * For invocation with β):
+   *   * Input Realms may be local or synced, encrypted or non-encrypted
+   *   * Output Realms will be local or synced, encrypted or non-encrypted, depending on the configuration passed to the function
    *
    * The destination file cannot already exist.
-   * When invoked on a synced realm, a copy of the realm is created that any user can open and
-   * resume synchronization with the server.
    *
    * Note that if this method is called from within a write transaction, the current data is written,
    * not the data from the point when the previous write transaction was committed.
-   * @param {string} path path to save the Realm to.
-   * @param {ArrayBuffer|ArrayBufferView} [encryptionKey] - Optional 64-byte encryption key to encrypt the new file with.
+   * @param {string|Realm~Configuration} pathOrConfig a α) path to save the Realm to, OR β) {@link Realm~Configuration | Configuration} that describes the output realm.
+   * @param {ArrayBuffer|ArrayBufferView} [encryptionKey] - Optional 64-byte encryption key to encrypt the new file with.  Must not be present when
+   * β) a {@link Realm~Configuration | Configuration} is given as first parameter, in which case encryptionKey can be set in {@link Realm~Configuration#encryptionKey}.
    */
-  writeCopyTo(path, encryptionKey) {}
+  writeCopyTo(pathOrConfig, encryptionKey) {}
 
   /**
    * Get the current schema version of the Realm at the given path.
@@ -378,6 +387,10 @@ class Realm {
  *     - `usedSize` - The total bytes used by data in the file.
  *   It returns `true` to indicate that an attempt to compact the file should be made. The compaction
  *   will be skipped if another process is accessing it.
+ * @property {callback(realm)} [onFirstOpen] - The function called when opening a Realm for
+ *   the first time. The function can populate the Realm prior to opening it. When calling the callback,
+ *   the Realm will be in a write transaction. The callback takes one argument:
+ *     - `realm` - The Realm
  * @property {string} [path={@link Realm.defaultPath}] - The path to the file where the
  *   Realm database should be stored.
  * @property {string} [fifoFilesFallbackPath] - Opening a Realm creates a number of FIFO special files in order to