+++ templates/react-native-template-realm-js/template/App.js
@@ -1,21 +1,22 @@
-import React, { useCallback, useMemo } from "react";
-import { SafeAreaView, View, StyleSheet } from "react-native";
+import React, {useCallback, useMemo} from 'react';
+import {SafeAreaView, View, StyleSheet} from 'react-native';
 
-import TaskContext, { Task } from "./app/models/Task";
-import IntroText from "./app/components/IntroText";
-import AddTaskForm from "./app/components/AddTaskForm";
-import TaskList from "./app/components/TaskList";
-import colors from "./app/styles/colors";
+import TaskContext, {Task} from './app/models/Task';
+import IntroText from './app/components/IntroText';
+import AddTaskForm from './app/components/AddTaskForm';
+import TaskList from './app/components/TaskList';
+import colors from './app/styles/colors';
 
-const { useRealm, useQuery, RealmProvider } = TaskContext;
+const {useRealm, useQuery, RealmProvider} = TaskContext;
 
 function App() {
   const realm = useRealm();
-  const result = useQuery("Task");
-  const tasks = useMemo(() => result.sorted("createdAt"), [result]);
+  const result = useQuery(Task);
+
+  const tasks = useMemo(() => result.sorted('createdAt'), [result]);
 
   const handleAddTask = useCallback(
-    (description) => {
+    description => {
       if (!description) {
         return;
       }
@@ -28,14 +29,14 @@ function App() {
       // of sync participants to successfully sync everything in the transaction, otherwise
       // no changes propagate and the transaction needs to start over when connectivity allows.
       realm.write(() => {
-        realm.create("Task", new Task({description}));
+        realm.create('Task', Task.generate(description));
       });
     },
     [realm],
   );
 
   const handleToggleTaskStatus = useCallback(
-    (task) => {
+    task => {
       realm.write(() => {
         // Normally when updating a record in a NoSQL or SQL database, we have to type
         // a statement that will later be interpreted and used as instructions for how
@@ -59,7 +60,7 @@ function App() {
   );
 
   const handleDeleteTask = useCallback(
-    (task) => {
+    task => {
       realm.write(() => {
         realm.delete(task);
 
@@ -77,7 +78,11 @@ function App() {
         {tasks.length === 0 ? (
           <IntroText />
         ) : (
-          <TaskList tasks={tasks} onToggleTaskStatus={handleToggleTaskStatus} onDeleteTask={handleDeleteTask} />
+          <TaskList
+            tasks={tasks}
+            onToggleTaskStatus={handleToggleTaskStatus}
+            onDeleteTask={handleDeleteTask}
+          />
         )}
       </View>
     </SafeAreaView>