+++ templates/react-native-template-realm-ts/template/index.js
@@ -4,7 +4,7 @@
 
 import 'react-native-get-random-values';
 import {AppRegistry} from 'react-native';
-import App from './App';
+import App from './app/App';
 import {name as appName} from './app.json';
 
 AppRegistry.registerComponent(appName, () => App);